learnings/02_chatbot_with_memory_in_txt_file/pyproject.toml
[project]
name = "02-chatbot-with-memory-in-txt-file"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.2",
    "langchain>=1.2.13",
    "langchain-openai>=1.1.12",
    "langgraph>=1.1.3",
    "langgraph-checkpoint-postgres>=3.0.5",
    "psycopg-binary>=3.3.3",
    "psycopg-pool>=3.3.0",
    "pydantic-settings>=2.13.1",
    "uvicorn>=0.42.0",
]
